Implemented allocator across all backends (only used in CUDA at the moment) Templated Cuda backend and utils on "using Unified Memory or not" in prep for Vulkan memory Changed per-dimension size type to uint32_t in non-vulkan scenarios

+++ project/framework_test/src/cmd_line/subapps/MakeInputSubApp.cpp
@@ -28,7 +28,7 @@ void MakeInputSubApp::run() {
     }
 
     auto size = SimSize(
-            {(size_t) width-2, (size_t) height-2},
+            {(uint32_t) width-2, (uint32_t) height-2},
             physicalSize
     );
     auto simSnapshot = SimSnapshot(size);
